+++ src/ProjectReference.php
@@ -2,7 +2,6 @@
 
 namespace NumNum\UBL;
 
-use DateTime;
 use Sabre\Xml\Writer;
 use Sabre\Xml\XmlSerializable;